refactor to separate build and runtime components
