--- python/pyraphtory_jvm/pyraphtory_jvm/data/ivys/.gitignore
@@ -1,3 +1,0 @@
-*
-*/
-!.gitignore
