--- python/pyraphtory_jvm/pyraphtory_jvm/data/lib/.gitignore
@@ -1,3 +1,0 @@
-*
-*/
-!.gitignore
