+++ python/pyraphtory_jvm/setup.cfg
@@ -1,0 +1,8 @@
+[options]
+packages = find_namespace:
+package_dir =
+    =src
+include_package_data = True
+
+[options.packages.find]
+where = src
